[refactor] Dto Nested 구조 공통 dto  분리 - #26
